Make UTF conversion functions public
