+++ doc/changelog.dox
@@ -14,6 +14,7 @@
 
 - Require C++11 compatible compiler and stdlib
 - LFS: Add support for files > 2 GB where the underlying system supports it
+- Generic UTF conversion functions are now available in the boost::nowide::utf namespace
 
 \subsection changelog_10_0_2 Nowide 10.0.2
