+++ include/boost/nowide/stackstring.hpp
@@ -9,6 +9,7 @@
 #define BOOST_NOWIDE_STACKSTRING_HPP_INCLUDED
 
 #include <boost/nowide/convert.hpp>
+#include <boost/nowide/utf/utf.hpp>
 #include <cassert>
 #include <cstring>
 
@@ -91,7 +92,7 @@ namespace nowide {
         output_char* convert(const input_char* input)
         {
             if(input)
-                return convert(input, input + detail::strlen(input));
+                return convert(input, input + utf::strlen(input));
             clear();
             return get();
         }
@@ -107,15 +108,15 @@ namespace nowide {
                 // Minimum size required: 1 output char per input char + trailing NULL
                 const size_t min_output_size = input_len + 1;
                 // If there is a chance the converted string fits on stack, try it
-                if(min_output_size <= buffer_size && detail::convert_buffer(buffer_, buffer_size, begin, end))
+                if(min_output_size <= buffer_size && utf::convert_buffer(buffer_, buffer_size, begin, end))
                     data_ = buffer_;
                 else
                 {
                     // Fallback: Allocate a buffer that is surely large enough on heap
                     // Max size: Every input char is transcoded to the output char with maximum with + trailing NULL
-                    const size_t max_output_size = input_len * detail::utf::utf_traits<output_char>::max_width + 1;
+                    const size_t max_output_size = input_len * utf::utf_traits<output_char>::max_width + 1;
                     data_ = new output_char[max_output_size];
-                    const bool success = detail::convert_buffer(data_, max_output_size, begin, end) == data_;
+                    const bool success = utf::convert_buffer(data_, max_output_size, begin, end) == data_;
                     assert(success);
                     (void)success;
                 }
